Move window managment code from Scene to Window
